pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vigil-cv"
version = "0.1.0"
description = "Edge anomaly detection for video streams with multi-backend inference benchmarking"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Traelyn"},
]

dependencies = [
    "torch>=2.3",
    "torchvision>=0.18",
    "onnx>=1.16",
    "onnxruntime-gpu>=1.18",
    "onnxsim>=0.4",
    "numpy>=1.26",
    "scipy>=1.12",
    "scikit-learn>=1.4",
    "opencv-python-headless>=4.9",
    "gradio>=4.0",
    "tqdm>=4.66",
    "rich>=13.0",
    "matplotlib>=3.8",
    "pandas>=2.2",
    "faiss-cpu>=1.8",
]

[project.optional-dependencies]
gpu = [
    "faiss-gpu>=1.8",
]
tensorrt = [
    # TensorRT 10.x on Windows must be installed from the NVIDIA zip package.
    # See README.md "TensorRT on Windows" section for instructions.
    # "tensorrt>=10.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.setuptools.packages.find]
include = ["data*", "models*", "training*", "export*", "inference*", "benchmark*", "demo*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
